requirements.txt
python-telegram-bot==20.7
requests==2.31.0
aiohttp==3.8.6